elevator v0.1.0 NetworkHandler

Module for handling and setting up the Node Network, and communication between nodes

State:

A map with the connected nodes as keys and a backup of the node's elevator state as the respective value.

Tasks:

Initializes all modules for one computer, broadcasts own IP and listnes, making a Peer-to-peer network of NetworkModules

Decides the recipient of Hall orders based on a cost function, considering number of orders and distance to order.

Is responsible for restarting nodes and redistributing orders that are not executed by assigned elevator

Communication:

Receives from: OrderHandler, WatchDog, (other nodes') NetworkModule(s)

Sends to: OrderHandler, (other nodes') NetworkModule(s)

Functions

all_nodes()

Returns all nodes in the current cluster. Returns a list of nodes or an error message

Examples

iex> NetworkStuff.all_nodes
[:'heis@10.100.23.253', :'heis@10.100.23.226']
iex> NetworkStuff.all_nodes
{:error, :node_not_running}

boot_node(node_name, tick_time \\ 15000)

boots a node with a specified tick time. node_name sets the node name before @. The IP-address is automatically imported

iex> NetworkStuff.boot_node "frank"
{:ok, #PID<0.12.2>}
iex(frank@10.100.23.253)> _

broadcast_self(socket, recv_PORT, name)

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

cost_function(state, order)

export_order(arg)

get_my_ip()

Returns (hopefully) the ip address of your network interface.

Examples

iex> NetworkStuff.get_my_ip
{10, 100, 23, 253}

handle_cast(msg, state)

An elevator is chosen for the specific order, using the cost function

init(list)

Boot Node with name "elev@ip" and spawn listen and receive processes based on UDP

ip_to_string(ip)

formats an ip address on tuple format to a bytestring

Examples

iex> NetworkStuff.ip_to_string {10, 100, 23, 253}
'10.100.23.253'

listen(socket, network_handler_pid)

multi_call_request_order_rank(order)

multi_call_update_backup(backup)

start_link(list \\ [20088, 20086])

sync(ext_order_list)

synchronize_order_lists(order_list)

test()