Elevator v1.0.0 OrderHandler

Module for receiving and handling orders form Poller and NetworkHandler and synchronizing orders

State:

A list of order structs, where the hall orders are shared between OrderHandlers and cab orders are module specific

Tasks:

Keeps track of orders relevant for the specific node

Processes orders received from Poller and Network Handler

Communication:

Sends to: StateMachine, NetworkHandler

Receives from: StateMachine, NetworkHandler, Poller

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

distribute_order(order, chosen \\ false)

distribute order from Poller or NetworkHandler, sending hall orders to NetworkHandler if it has not been chosen and passing the other orders to own elevator

handle_cast(msg, state)

Synchronize the order list with an external one

init(order_list \\ [])

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

start_link(init_args)

sync_order(order_list)

synchronize non-cab orders to other nodes' order lists