Elevator v1.0.0 Overseer

A Supervisor that keeps track and restarts entire program upon request or crash, using the one_for_all strategy, as well as being the entry point for the entire program

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(list)

Initialization starting all modules under a Supervisor with one_for_all strategy

start_link()