Elevator v1.0.0 Poller

Module for polling the DriverInterface to the hardware and passing the information

State:

No state

Tasks:

Spawning two threads polling the floor sensor and button sensor, and passing these messages to the OrderHandler and StateMachine

Communication:

Sends to: OrderHandler, StateMachine

Receives from:

Functions

button_poller()

Loop and send when a new button is pressed 20 times per second

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

floor_poller(floor)

Loop and send when a new floor is reached 20 times per second

init(mock)

Initialize by turning off all lights and spawning one process for button polling and one for floor polling

send_floor(floor)

set_order(floor, button_type)

start_link(init_args)