Elevator v0.1.0 NetworkHandler

Module for handling and setting up the Node Network, and communication between nodes

State:

A map with the connected nodes as keys and a backup of the node's elevator state as the respective value, as well as a boolean value indicating their ability to accept new orders. The entire state for three connected nodes will look like this: %{ "elev@10.100.23.197": [ %State{active_orders: [], direction: :stop, floor: 0}, true ], "elev@10.100.23.233": [ %State{active_orders: [], direction: :stop, floor: 1}, true ], "elev@10.100.23.253": [ %State{active_orders: [], direction: :stop, floor: 1}, true ] }

Tasks:

Initializes all modules for one computer, broadcasts own IP and listnes, making a Peer-to-peer network of NetworkModules

Decides the recipient of Hall orders based on a cost function, considering number of orders and distance to order.

Is responsible for restarting nodes and redistributing orders that are not executed by assigned elevator

Communication:

Receives from: OrderHandler, WatchDog, StateMachine, (other nodes') NetworkModule(s)

Sends to: OrderHandler, (other nodes') NetworkModule(s)

Functions

all_nodes()

Courtesy of @jostlowe Returns all nodes in the current cluster. Returns a list of nodes or an error message

Examples

iex> NetworkStuff.all_nodes
[:'heis@10.100.23.253', :'heis@10.100.23.226']
iex> NetworkStuff.all_nodes
{:error, :node_not_running}

boot_node(node_name, tick_time \\ 15000)

Courtesy of @jostlowe boots a node with a specified tick time. node_name sets the node name before @. The IP-address is automatically imported

iex> NetworkStuff.boot_node "frank"
{:ok, #PID<0.12.2>}
iex(frank@10.100.23.253)> _

broadcast_self(socket, recv_port, name)

Continously broadcasting own name with UDP

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

cost_function(state, order)

Cost function for assigning a cost for an elevator to execute an order, considering both the distance to the order and the number of orders already assigned

distance_to_order(elevator_order, elevator_state)

Calculates the distance to the order Example iex> distance_to_order(%Order{floor: 1, type: hall_up}, %State{floor: 3, direction: :up, active_orders: []}) 2

export_order(arg)

Send assigned order to other node

get_my_ip()

Courtesy of @jostlowe Returns (hopefully) the ip address of your network interface.

Examples

iex> NetworkStuff.get_my_ip
{10, 100, 23, 253}

handle_call(msg, from, state)

Handle call for receiving a order from external node to be assigned to own state machine

handle_cast(msg, state)

Handle an order cast internally on an elevator

handle_info(msg, state)

Handle event of node going down.

init(list)

Boot Node with name "elev@ip" and spawn listen and receive processes based on UDP broadcasting

ip_to_string(ip)

Courtesy of @jostlowe formats an ip address on tuple format to a bytestring

Examples

iex> NetworkStuff.ip_to_string {10, 100, 23, 253}
'10.100.23.253'

listen(socket, network_handler_pid)

Continously listen for messages broadcasted, request connection to the node with the name received

monitor_me_back(node_name)

Set up two-way monitoring

multi_call_request_backup(from_node_name, about_node)

Get info about a node in another node's state map

multi_call_update_backup(backup)

Send own updated backup to all other nodes

redistribute_orders(order_list)

Iterate through an order list and send all orders to own state machine

return_cab_orders(node_name, net_state)

Process to be spawned for delaying cab order resend to after initialization. Example iex> pid = Process.spawn(NetworkHandler, :resend_cab_orders, [node_name, net_state], []) iex> Process.send_after(pid, :resend_cab_orders, 3000)

start_link(list \\ [20087, 20086])

sync(ext_order_list)

Synchronize local order list with an external order list

synchronize_order_lists(order_list)

Synchronize internal order list with the other elevators